Adding: Adding image Image Browser as a submodule Fixing: Minor cleaning up code
Well the long awaited Image browser is no include. Eventhough we have a feature freeze due the RC, i added it. It is eagerly awaited and as it implements itself as a submodule and i did not neeed to add any code to the core module, i was happy to include ist.
The browser was mainly developed by Cvelde with only a little help of mine. So koodos to him! Thank you Cvelde!