Added: http://drupal.org/node/748342 Fixed: http://drupal.org/node/746974
With this version you can finally add images to comment-forms. Miro helped me a lot on this task, took as suprisinly long though.
Big parts of the cache handling and file processing are reviewd and rewritten. So those are quite huge changes, eventhough the backend stays the same of course.
The dependecy to FAPI cacheable forms is no longer there (this was one of the tasks for the comment support).